Move tests to appripriate locations, use less pd.
