+++ pandas/tests/window/test_api.py
@@ -1,8 +1,6 @@
 import numpy as np
 import pytest
 
-import pandas.util._test_decorators as td
-
 import pandas as pd
 from pandas import DataFrame, Index, Series, Timestamp, concat
 import pandas._testing as tm
@@ -235,30 +233,6 @@ def test_count_nonnumeric_types():
 
     result = df.rolling(1, min_periods=0).count()
     expected = df.notna().astype(float)
-    tm.assert_frame_equal(result, expected)
-
-
-@td.skip_if_no_scipy
-@pytest.mark.filterwarnings("ignore:can't resolve:ImportWarning")
-def test_window_with_args():
-    # make sure that we are aggregating window functions correctly with arg
-    r = Series(np.random.randn(100)).rolling(
-        window=10, min_periods=1, win_type="gaussian"
-    )
-    expected = concat([r.mean(std=10), r.mean(std=0.01)], axis=1)
-    expected.columns = ["<lambda>", "<lambda>"]
-    result = r.aggregate([lambda x: x.mean(std=10), lambda x: x.mean(std=0.01)])
-    tm.assert_frame_equal(result, expected)
-
-    def a(x):
-        return x.mean(std=10)
-
-    def b(x):
-        return x.mean(std=0.01)
-
-    expected = concat([r.mean(std=10), r.mean(std=0.01)], axis=1)
-    expected.columns = ["a", "b"]
-    result = r.aggregate([a, b])
     tm.assert_frame_equal(result, expected)
