+++ pandas/tests/window/test_apply.py
@@ -1,9 +1,6 @@
 import numpy as np
 import pytest
 
-from pandas.errors import NumbaUtilError
-import pandas.util._test_decorators as td
-
 from pandas import DataFrame, Index, MultiIndex, Series, Timestamp, date_range
 import pandas._testing as tm
 
@@ -131,14 +128,6 @@ def test_invalid_raw_numba():
         ValueError, match="raw must be `True` when using the numba engine"
     ):
         Series(range(1)).rolling(1).apply(lambda x: x, raw=False, engine="numba")
-
-
-@td.skip_if_no("numba")
-def test_invalid_kwargs_nopython():
-    with pytest.raises(NumbaUtilError, match="numba does not support kwargs with"):
-        Series(range(1)).rolling(1).apply(
-            lambda x: x, kwargs={"a": 1}, engine="numba", raw=True
-        )
 
 
 @pytest.mark.parametrize("args_kwargs", [[None, {"par": 10}], [(10,), None]])
