+++ pandas/tests/window/test_expanding.py
@@ -118,30 +118,27 @@ def test_expanding_axis(axis_frame):
     tm.assert_frame_equal(result, expected)
 
 
-@pytest.mark.parametrize("constructor", [Series, DataFrame])
-def test_expanding_count_with_min_periods(constructor):
+def test_expanding_count_with_min_periods(frame_or_series):
     # GH 26996
-    result = constructor(range(5)).expanding(min_periods=3).count()
-    expected = constructor([np.nan, np.nan, 3.0, 4.0, 5.0])
+    result = frame_or_series(range(5)).expanding(min_periods=3).count()
+    expected = frame_or_series([np.nan, np.nan, 3.0, 4.0, 5.0])
     tm.assert_equal(result, expected)
 
 
-@pytest.mark.parametrize("constructor", [Series, DataFrame])
-def test_expanding_count_default_min_periods_with_null_values(constructor):
+def test_expanding_count_default_min_periods_with_null_values(frame_or_series):
     # GH 26996
     values = [1, 2, 3, np.nan, 4, 5, 6]
     expected_counts = [1.0, 2.0, 3.0, 3.0, 4.0, 5.0, 6.0]
 
-    result = constructor(values).expanding().count()
-    expected = constructor(expected_counts)
+    result = frame_or_series(values).expanding().count()
+    expected = frame_or_series(expected_counts)
     tm.assert_equal(result, expected)
 
 
-@pytest.mark.parametrize("constructor", [Series, DataFrame])
-def test_expanding_count_with_min_periods_exceeding_series_length(constructor):
+def test_expanding_count_with_min_periods_exceeding_series_length(frame_or_series):
     # GH 25857
-    result = constructor(range(5)).expanding(min_periods=6).count()
-    expected = constructor([np.nan, np.nan, np.nan, np.nan, np.nan])
+    result = frame_or_series(range(5)).expanding(min_periods=6).count()
+    expected = frame_or_series([np.nan, np.nan, np.nan, np.nan, np.nan])
     tm.assert_equal(result, expected)
 
 
@@ -246,10 +243,9 @@ def test_center_deprecate_warning():
         df.expanding()
 
 
-@pytest.mark.parametrize("constructor", ["DataFrame", "Series"])
-def test_expanding_sem(constructor):
+def test_expanding_sem(frame_or_series):
     # GH: 26476
-    obj = getattr(pd, constructor)([0, 1, 2])
+    obj = frame_or_series([0, 1, 2])
     result = obj.expanding().sem()
     if isinstance(result, DataFrame):
         result = Series(result[0].values)
